+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [8.57.0] - 2019-06-27
+
+### Changed
+
+- **Table** new line button with action replaced by a **ButtonGroup**.
+
+### Removed
+
+- **EXPERIMENTAL_ButtonWithAction** in favor of ButtonGroup.
+
 ## [8.56.0] - 2019-06-27
 
 ### Added